refactor: combined student and lecturer login form
