+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^17.0.2",
     "@types/socket.io-client": "^1.4.36",
     "bootstrap": "^4.6.0",
+    "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.1",
