+++ frontend/src/components/LoginForm.tsx
@@ -1,25 +1,21 @@
 import { Form, Button } from 'react-bootstrap';
-import { useHistory, Link } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
-const StudentLogin = () => {
-  let history = useHistory();
-
+const LecturerLogin = (props: { isLecturer: boolean }) => {
   const handleSubmit = (event: any) => {
     event.preventDefault();
     console.log(event.target.email.value);
     console.log(event.target.password.value);
 
     // login logic
-    
+
     event.target.email.value = '';
     event.target.password.value = '';
-
-    history.push("/student/dashboard");
   };
 
   return (
     <div className="container-fluid vh-100 d-flex flex-column justify-content-center align-items-center p-2 bg-light">
-      <h1 className="mb-5">hello fellow student!</h1>
+      <h1 className="mb-5">hello fellow {props.isLecturer ? 'lecturer' : 'student'}!</h1>
       <Form onSubmit={handleSubmit}>
         <Form.Group controlId="email">
           <Form.Label>Email address</Form.Label>
@@ -33,20 +29,24 @@ const StudentLogin = () => {
           <Form.Label>Password</Form.Label>
           <Form.Control type="password" placeholder="Password" />
         </Form.Group>
-        <Button variant="primary" type="submit" block className="mb-5">
+        <Button variant="primary" type="submit" block>
           sign in
         </Button>
-        <Form.Group>
-          <Form.Text className="text-muted text-center">
-            do not have an account yet?
-          </Form.Text>
-          <Link to="/register/student">
-            <button type="button" className="btn btn-outline-primary btn-block">sign up</button>
-          </Link>
-        </Form.Group>
+        {!props.isLecturer ?
+          <Form.Group className="mt-5">
+            <Form.Text className="text-muted text-center">
+              do not have an account yet?
+            </Form.Text>
+            <Link to="/register/student">
+              <button type="button" className="btn btn-outline-primary btn-block">sign up</button>
+            </Link>
+          </Form.Group>
+          : null
+        }
+
       </Form>
     </div>
   );
 };
 
-export default StudentLogin;
+export default LecturerLogin;
